mcp_python/requirements.txt
mcp>=1.0.0
pytesseract>=0.3.10
Pillow>=10.0.0
